second commit

+++ src/main/java/org/islom/homework212/repository/LanguageSupportRepo.java
@@ -3,5 +3,5 @@ package org.islom.homework212.repository;
 import org.islom.homework212.entity.LanguageSupport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LanguageSupportRepo extends JpaRepository<LanguageSupport, Long> {
+public interface LanguageSupportRepo extends JpaRepository<LanguageSupport, Integer> {
 }
