+++ src/main/java/org/islom/homework212/repository/MissCodeContentRepo.java
@@ -3,5 +3,5 @@ package org.islom.homework212.repository;
 import org.islom.homework212.entity.MissCodeContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MissCodeContentRepo extends JpaRepository<MissCodeContent, Long> {
+public interface MissCodeContentRepo extends JpaRepository<MissCodeContent, Integer> {
 }
